requirements.txt
numpy
opencv-python
matplotlib
ffmpeg-python
colorlog